refactor: 말 초기 생성 로직 다른 클래스로 분리
