+++ src/main/java/chess/board/ChessBoard.java
@@ -1,14 +1,7 @@
 package chess.board;
 
-import chess.piece.Bishop;
 import chess.piece.EmptyPiece;
-import chess.piece.King;
-import chess.piece.Knight;
-import chess.piece.Pawn;
 import chess.piece.Piece;
-import chess.piece.Queen;
-import chess.piece.Rook;
-import chess.piece.Team;
 
 import java.util.Collections;
 import java.util.HashMap;
@@ -23,15 +16,7 @@ public class ChessBoard {
 
     public static ChessBoard createBoard() {
         final Map<Position, Piece> piecePosition = new HashMap<>();
-        initPosition(piecePosition);
-        for (final Team team : Team.values()) {
-            createPawn(piecePosition, Rank.from(team, true), team);
-            createRook(piecePosition, Rank.from(team, false), team);
-            createKnight(piecePosition, Rank.from(team, false), team);
-            createBishop(piecePosition, Rank.from(team, false), team);
-            createQueen(piecePosition, Rank.from(team, false), team);
-            createKing(piecePosition, Rank.from(team, false), team);
-        }
+        PieceFactory.createPiece(piecePosition);
         return new ChessBoard(piecePosition);
     }
 
@@ -41,14 +26,6 @@ public class ChessBoard {
 
     private ChessBoard(final Map<Position, Piece> piecePosition) {
         this.piecePosition = piecePosition;
-    }
-
-    private static void initPosition(final Map<Position, Piece> piecePosition) {
-        for (final File file : File.values()) {
-            for (final Rank rank : Rank.values()) {
-                piecePosition.put(new Position(file, rank), new EmptyPiece());
-            }
-        }
     }
 
     public void movePiece(final Position from, final Position to) {
@@ -161,7 +138,7 @@ public class ChessBoard {
         final int max = Math.max(fromFile.getIndex(), toFile.getIndex()) - 1;
 
         for (int i = min; i <= max; i++) {
-            Piece validationPiece = piecePosition.get(new Position(File.of(i), from.getRank()));
+            Piece validationPiece = piecePosition.get(new Position(File.from(i), from.getRank()));
             validateBlockedRoute(validationPiece);
         }
     }
@@ -170,35 +147,6 @@ public class ChessBoard {
         if (!validationPiece.isEmpty()) {
             throw new IllegalArgumentException("말이 이동경로에 존재하여 이동할 수 없습니다.");
         }
-    }
-
-    private static void createPawn(final Map<Position, Piece> piecePosition, final Rank rank, final Team team) {
-        for (final File file : File.values()) {
-            piecePosition.put(new Position(file, rank), new Pawn(team));
-        }
-    }
-
-    private static void createRook(final Map<Position, Piece> piecePosition, final Rank rank, final Team team) {
-        piecePosition.put(new Position(File.A, rank), new Rook(team));
-        piecePosition.put(new Position(File.H, rank), new Rook(team));
-    }
-
-    private static void createKnight(final Map<Position, Piece> piecePosition, final Rank rank, final Team team) {
-        piecePosition.put(new Position(File.B, rank), new Knight(team));
-        piecePosition.put(new Position(File.G, rank), new Knight(team));
-    }
-
-    private static void createBishop(final Map<Position, Piece> piecePosition, final Rank rank, final Team team) {
-        piecePosition.put(new Position(File.C, rank), new Bishop(team));
-        piecePosition.put(new Position(File.F, rank), new Bishop(team));
-    }
-
-    private static void createQueen(final Map<Position, Piece> piecePosition, final Rank rank, final Team team) {
-        piecePosition.put(new Position(File.D, rank), new Queen(team));
-    }
-
-    private static void createKing(final Map<Position, Piece> piecePosition, final Rank rank, final Team team) {
-        piecePosition.put(new Position(File.E, rank), new King(team));
     }
 
     public Map<Position, Piece> getPiecePosition() {
